requirements.txt
# Telegram Bot API
python-telegram-bot==21.3

# HTTP-клиент для Hugging Face API
httpx==0.27.0

# Переменные окружения из .env
python-dotenv==1.0.1